Update employees-vehicles-list.component.html to include tabs for employees and vehicles
This commit updates the employees-vehicles-list.component.html file to include tabs for displaying employees and vehicles. The tab component is implemented using an unordered list with two list items, one for employees and one for vehicles. Each list item contains an SVG icon and a text label. The active tab is determined by the value of the tabType property. The commit also includes changes to the filter button and add item button to accommodate the new tab layout.